Renaming, refactoring, and removing some of the vector jmh benchmarks
