+++ benchmarks/src/main/java/org/elasticsearch/benchmark/vector/scorer/VectorScorerDistanceBulkBenchmark.java
@@ -6,7 +6,7 @@
  * your election, the "Elastic License 2.0", the "GNU Affero General Public
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
-package org.elasticsearch.benchmark.vector;
+package org.elasticsearch.benchmark.vector.scorer;
 
 import org.apache.lucene.util.VectorUtil;
 import org.apache.lucene.util.quantization.OptimizedScalarQuantizer;
@@ -38,7 +38,7 @@ import java.util.concurrent.TimeUnit;
 @Measurement(iterations = 5, time = 1)
 // engage some noise reduction
 @Fork(value = 1)
-public class DistanceBulkBenchmark {
+public class VectorScorerDistanceBulkBenchmark {
 
     static {
         LogConfigurator.configureESLogging(); // native access requires logging to be initialized
