+++ benchmarks/src/main/java/org/elasticsearch/benchmark/vector/scorer/VectorScorerJDKInt7uBenchmark.java
@@ -6,7 +6,7 @@
  * your election, the "Elastic License 2.0", the "GNU Affero General Public
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
-package org.elasticsearch.benchmark.vector;
+package org.elasticsearch.benchmark.vector.scorer;
 
 import org.apache.lucene.util.VectorUtil;
 import org.elasticsearch.common.logging.LogConfigurator;
@@ -37,7 +37,7 @@ import java.util.concurrent.TimeUnit;
 @State(Scope.Benchmark)
 @Warmup(iterations = 3, time = 1)
 @Measurement(iterations = 5, time = 1)
-public class JDKVectorInt7uBenchmark {
+public class VectorScorerJDKInt7uBenchmark {
 
     static {
         NodeNamePatternConverter.setGlobalNodeName("foo");
